rename `client` mod -> `engine`
